+++ admin/src/main/resources/io/buoyant/admin/js/template/compiled_templates.js
@@ -111,19 +111,21 @@ templates['delegatenode'] = template({"1":function(container,depth0,helpers,part
 
   return "  <li class='list-group-item\n"
     + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.isPrimary : depth0),{"name":"if","hash":{},"fn":container.program(30, data, 0),"inverse":container.program(35, data, 0),"data":data})) != null ? stack1 : "")
-    + "'\n"
+    + "'\n\n"
     + ((stack1 = helpers["with"].call(alias1,(depth0 != null ? depth0.dentry : depth0),{"name":"with","hash":{},"fn":container.program(37, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
-    + ">\n    <div class='node-info'>\n"
+    + ">\n\n    <div class='node-info'>\n"
     + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.isNeg : depth0),{"name":"if","hash":{},"fn":container.program(39, data, 0),"inverse":container.program(41, data, 0),"data":data})) != null ? stack1 : "")
     + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.weight : depth0),{"name":"if","hash":{},"fn":container.program(52, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
-    + "    </div>\n"
+    + "    </div>\n\n"
     + ((stack1 = helpers["with"].call(alias1,(depth0 != null ? depth0.addr : depth0),{"name":"with","hash":{},"fn":container.program(54, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
-    + "    "
+    + "\n    <span>\n      "
+    + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.id : depth0),{"name":"if","hash":{},"fn":container.program(57, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
+    + "\n      <span class='path'>(residual: "
     + container.escapeExpression(((helper = (helper = helpers.path || (depth0 != null ? depth0.path : depth0)) != null ? helper : helpers.helperMissing),(typeof helper === "function" ? helper.call(alias1,{"name":"path","hash":{},"data":data}) : helper)))
-    + "\n"
-    + ((stack1 = helpers["with"].call(alias1,(depth0 != null ? depth0.dentry : depth0),{"name":"with","hash":{},"fn":container.program(57, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
+    + ")</span>\n    </span>\n\n"
+    + ((stack1 = helpers["with"].call(alias1,(depth0 != null ? depth0.dentry : depth0),{"name":"with","hash":{},"fn":container.program(59, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
     + "  </li>\n"
-    + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.transformed : depth0),{"name":"if","hash":{},"fn":container.program(59, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "");
+    + ((stack1 = helpers["if"].call(alias1,(depth0 != null ? depth0.transformed : depth0),{"name":"if","hash":{},"fn":container.program(61, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "");
 },"30":function(container,depth0,helpers,partials,data) {
     var stack1;
 
@@ -187,6 +189,12 @@ templates['delegatenode'] = template({"1":function(container,depth0,helpers,part
     + alias4(((helper = (helper = helpers.port || (depth0 != null ? depth0.port : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"port","hash":{},"data":data}) : helper)))
     + "\n";
 },"57":function(container,depth0,helpers,partials,data) {
+    var helper;
+
+  return "["
+    + container.escapeExpression(((helper = (helper = helpers.id || (depth0 != null ? depth0.id : depth0)) != null ? helper : helpers.helperMissing),(typeof helper === "function" ? helper.call(depth0 != null ? depth0 : {},{"name":"id","hash":{},"data":data}) : helper)))
+    + "]";
+},"59":function(container,depth0,helpers,partials,data) {
     var helper, alias1=depth0 != null ? depth0 : {}, alias2=helpers.helperMissing, alias3="function", alias4=container.escapeExpression;
 
   return "      <span class='node-dentry'>"
@@ -194,7 +202,7 @@ templates['delegatenode'] = template({"1":function(container,depth0,helpers,part
     + " => "
     + alias4(((helper = (helper = helpers.dst || (depth0 != null ? depth0.dst : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"dst","hash":{},"data":data}) : helper)))
     + "</span>\n";
-},"59":function(container,depth0,helpers,partials,data) {
+},"61":function(container,depth0,helpers,partials,data) {
     var stack1, helper;
 
   return "  <ul class='list-group'>"
@@ -206,7 +214,7 @@ templates['delegatenode'] = template({"1":function(container,depth0,helpers,part
   return ((stack1 = helpers["if"].call(depth0 != null ? depth0 : {},(depth0 != null ? depth0.child : depth0),{"name":"if","hash":{},"fn":container.program(1, data, 0),"inverse":container.program(29, data, 0),"data":data})) != null ? stack1 : "");
 },"useData":true});
 templates['delegator'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
-    return "<div class=\"row\">\n  <div class=\"col-lg-6\">\n    <div class=\"input-group path\">\n      <span class=\"input-group-addon\" id=\"basic-addon1\">Path</span>\n      <input type=\"text\" class=\"form-control\" id=\"path-input\" placeholder=\"/path/to/resource\" aria-describedby=\"basic-addon1\">\n      <span class=\"input-group-btn\">\n        <button class=\"btn btn-default go\" type=\"button\">Go!</button>\n      </span>\n    </div>\n  </div>\n</div>\n\n<div class=\"dtab\">\n  <h4 class=\"header\">\n    Dtab\n    <button id=\"edit-dtab-btn\" class=\"btn btn-info btn-sm\">Edit</button>\n    <button id=\"save-dtab-btn\" class=\"btn btn-success btn-sm hide disabled\">Apply*</button>\n  </h4>\n  <div id=\"dtab\"></div>\n  <div id=\"dtab-base\"></div>\n  <div id=\"dtab-edit\" class=\"hide\">\n    <textarea type=\"text\" class=\"form-control\" id=\"dtab-input\" placeholder=\"\"></textarea>\n  </div>\n  <div class=\"save-warning\">*Applies dtab for playground testing only!</div>\n</div>\n\n</div>\n<div class=\"result\">\n</div>\n<div class=\"modal fade error-modal\" tabindex=\"-1\" role=\"dialog\">\n</div>\n";
+    return "<div class=\"row\">\n  <div class=\"col-lg-6\">\n    <div class=\"input-group path\">\n      <span class=\"input-group-addon\" id=\"basic-addon1\">Path</span>\n      <input type=\"text\" class=\"form-control\" id=\"path-input\" placeholder=\"/path/to/resource\" aria-describedby=\"basic-addon1\">\n      <span class=\"input-group-btn\">\n        <button class=\"btn btn-default go\" type=\"button\">Go!</button>\n      </span>\n    </div>\n  </div>\n</div>\n\n<div class=\"dtab\">\n  <h4 class=\"header\">\n    Dtab\n  </h4>\n  <div id=\"dtab\"></div>\n  <div class=\"namerd-dtab-warning hide\">Note: the above Dentries are from namerd and can't be edited.</div>\n  <div id=\"dtab-base\"></div>\n\n  <div id=\"dtab-edit\" class=\"hide\">\n    <textarea type=\"text\" class=\"form-control\" id=\"dtab-input\" placeholder=\"\"></textarea>\n  </div>\n\n  <button id=\"edit-dtab-btn\" class=\"btn btn-info btn-sm\">Edit</button>\n  <button id=\"save-dtab-btn\" class=\"btn btn-success btn-sm hide disabled\">Apply*</button>\n  <a target=\"_blank\" id=\"reset-dtab-link\" class=\"hide\">Reset</a>\n  <div class=\"save-warning\">*Applies dtab for playground testing only!</div>\n</div>\n\n<div class=\"result\"></div>\n\n<div class=\"modal fade error-modal\" tabindex=\"-1\" role=\"dialog\"></div>\n\n<div class=\"modal fade confirm-modal\" tabindex=\"-1\" role=\"dialog\">\n  <div class=\"modal-dialog modal-sm\">\n    <div class=\"modal-content\">\n      <div class=\"modal-body\">Reset Dtab to initial state?</div>\n      <div class=\"modal-footer\">\n        <button type=\"button\" class=\"btn btn-secondary\" data-dismiss=\"modal\">Cancel</button>\n        <button type=\"button\" class=\"btn btn-warning confirm\" data-dismiss=\"modal\">Reset</button>\n      </div>\n    </div>\n  </div>\n</div>\n";
 },"useData":true});
 templates['dentry'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
     var helper, alias1=depth0 != null ? depth0 : {}, alias2=helpers.helperMissing, alias3="function", alias4=container.escapeExpression;
@@ -225,6 +233,23 @@ templates['error_modal'] = template({"compiler":[7,">= 4.0.0"],"main":function(c
     return "<div class=\"modal-dialog modal-sm\">\n  <div class=\"modal-content\">\n    <div class=\"modal-header\">\n      <button type=\"button\" class=\"close\" data-dismiss=\"modal\" aria-label=\"Close\"><span aria-hidden=\"true\">&times;</span></button>\n      <h4 class=\"modal-title\">Whoops!</h4>\n    </div>\n    <div class=\"modal-body\">\n      <p>Looks like there was an issue completing your request.</p>\n      <pre>"
     + container.escapeExpression(container.lambda(depth0, depth0))
     + "</pre>\n    </div>\n  </div>\n</div>\n";
+},"useData":true});
+templates['latencies.partial'] = template({"1":function(container,depth0,helpers,partials,data) {
+    var helper, alias1=depth0 != null ? depth0 : {}, alias2=helpers.helperMissing, alias3="function", alias4=container.escapeExpression;
+
+  return "      <div>\n        <span class=\"latency-label\">\n          <span class=\"latency-legend\" style=\"background-color:"
+    + alias4(((helper = (helper = helpers.latencyColor || (depth0 != null ? depth0.latencyColor : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyColor","hash":{},"data":data}) : helper)))
+    + ";\">&nbsp;</span>"
+    + alias4(((helper = (helper = helpers.latencyLabel || (depth0 != null ? depth0.latencyLabel : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyLabel","hash":{},"data":data}) : helper)))
+    + "\n        </span>\n        <span class=\"pull-right latency-value\">"
+    + alias4(((helper = (helper = helpers.latencyValue || (depth0 != null ? depth0.latencyValue : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyValue","hash":{},"data":data}) : helper)))
+    + " ms</span>\n      </div>\n";
+},"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
+    var stack1;
+
+  return "<div class=\"col-md-2 router-latencies-container\">\n  <div class=\"metric-header\">Latencies</div>\n  <div class=\"router-latencies\">\n"
+    + ((stack1 = helpers.each.call(depth0 != null ? depth0 : {},depth0,{"name":"each","hash":{},"fn":container.program(1, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
+    + "  </div>\n</div>\n";
 },"useData":true});
 templates['logging_row'] = template({"1":function(container,depth0,helpers,partials,data) {
     var stack1, alias1=container.lambda, alias2=container.escapeExpression;
@@ -363,17 +388,7 @@ templates['request_totals'] = template({"1":function(container,depth0,helpers,pa
     + ((stack1 = helpers.each.call(depth0 != null ? depth0 : {},(depth0 != null ? depth0.metrics : depth0),{"name":"each","hash":{},"fn":container.program(1, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
     + "</div>\n";
 },"useData":true});
-templates['router_client'] = template({"1":function(container,depth0,helpers,partials,data) {
-    var helper, alias1=depth0 != null ? depth0 : {}, alias2=helpers.helperMissing, alias3="function", alias4=container.escapeExpression;
-
-  return "        <div>\n          <span class=\"latency-label\">\n            <span class=\"latency-legend\" style=\"background-color:"
-    + alias4(((helper = (helper = helpers.latencyColor || (depth0 != null ? depth0.latencyColor : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyColor","hash":{},"data":data}) : helper)))
-    + ";\">&nbsp;</span>"
-    + alias4(((helper = (helper = helpers.latencyLabel || (depth0 != null ? depth0.latencyLabel : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyLabel","hash":{},"data":data}) : helper)))
-    + "\n          </span>\n          <span class=\"pull-right latency-value\">"
-    + alias4(((helper = (helper = helpers.latencyValue || (depth0 != null ? depth0.latencyValue : depth0)) != null ? helper : alias2),(typeof helper === alias3 ? helper.call(alias1,{"name":"latencyValue","hash":{},"data":data}) : helper)))
-    + " ms</span>\n        </div>\n";
-},"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
+templates['router_client'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
     var stack1;
 
   return "<div class=\"client-metrics row\">\n  <div class=\"col-md-2\">\n"
@@ -382,9 +397,9 @@ templates['router_client'] = template({"1":function(container,depth0,helpers,par
     + "  </div>\n\n  <div class=\"col-md-2\">\n"
     + ((stack1 = container.invokePartial(partials.metricPartial,((stack1 = (depth0 != null ? depth0.data : depth0)) != null ? stack1.connections : stack1),{"name":"metricPartial","hash":{"metricClass":"metric-large","containerClass":"metric-container"},"data":data,"indent":"    ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
     + ((stack1 = container.invokePartial(partials.metricPartial,((stack1 = (depth0 != null ? depth0.data : depth0)) != null ? stack1.failures : stack1),{"name":"metricPartial","hash":{"metricClass":"failure-metric metric-large","containerClass":"metric-container"},"data":data,"indent":"    ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
-    + "  </div>\n\n  <div class=\"col-md-2 router-latencies-container\">\n    <div class=\"metric-header\">Latencies</div>\n    <div class=\"router-latencies\">\n"
-    + ((stack1 = helpers.each.call(depth0 != null ? depth0 : {},(depth0 != null ? depth0.latencies : depth0),{"name":"each","hash":{},"fn":container.program(1, data, 0),"inverse":container.noop,"data":data})) != null ? stack1 : "")
-    + "    </div>\n  </div>\n</div>\n";
+    + "  </div>\n\n"
+    + ((stack1 = container.invokePartial(partials.latencyPartial,(depth0 != null ? depth0.latencies : depth0),{"name":"latencyPartial","data":data,"indent":"  ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
+    + "</div>\n";
 },"usePartial":true,"useData":true});
 templates['router_client_container'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
     var helper, alias1=depth0 != null ? depth0 : {}, alias2=helpers.helperMissing, alias3="function", alias4=container.escapeExpression;
@@ -402,7 +417,7 @@ templates['router_container'] = template({"1":function(container,depth0,helpers,
     + alias2(alias1(depth0, depth0))
     + " row\" data-router=\""
     + alias2(alias1(depth0, depth0))
-    + "\">\n    <div class=\"summary row\"></div>\n\n    <div class=\"combined-client-graph\">\n      <div class=\"router-graph-header\">Requests per client</div>\n      <canvas class=\"router-graph\" height=\"181\"></canvas>\n    </div>\n\n    <div class=\"router-stats row\">\n      <div class=\"retries-bar-chart col-md-6\"></div>\n      <div class=\"retries-stats col-md-6\"></div>\n      <div class=\"clearfix\"></div>\n    </div>\n\n    <div class=\"clients router-clients\">\n      <div class=\"router-subsection-title\">Clients</div>\n    </div>\n    <div class=\"servers router-servers row\">\n      <div class=\"router-subsection-title\">Servers</div>\n    </div>\n  </div>\n";
+    + "\">\n    <div class=\"summary row\"></div>\n\n    <div class=\"combined-client-graph\">\n      <div class=\"router-graph-header\">Requests per client</div>\n      <canvas class=\"router-graph\" height=\"181\"></canvas>\n    </div>\n\n    <div class=\"router-stats row\">\n      <div class=\"retries-bar-chart col-md-6\"></div>\n      <div class=\"retries-stats col-md-6\"></div>\n      <div class=\"clearfix\"></div>\n    </div>\n\n    <div class=\"clients router-clients\">\n      <div class=\"pull-left router-subsection-title\">Clients</div>\n      <div class=\"pull-right client-toggle\">\n        <a target=\"blank\" class=\"expand-all\">expand all</a> &middot;\n        <a target=\"blank\" class=\"collapse-all\">collapse all</a>\n      </div>\n    </div>\n    <div class=\"servers router-servers row\">\n      <div class=\"router-subsection-title\">Servers</div>\n    </div>\n  </div>\n";
 },"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
     var stack1;
 
@@ -424,18 +439,23 @@ templates['router_option'] = template({"1":function(container,depth0,helpers,par
     + "</a></li>\n";
 },"useData":true});
 templates['router_server'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
-    var stack1, helper;
+    var stack1;
 
-  return "<div class=\"server-header router-header-large\">\n  "
-    + container.escapeExpression(((helper = (helper = helpers.server || (depth0 != null ? depth0.server : depth0)) != null ? helper : helpers.helperMissing),(typeof helper === "function" ? helper.call(depth0 != null ? depth0 : {},{"name":"server","hash":{},"data":data}) : helper)))
-    + "\n</div>\n\n<div class=\"container-fluid clearfix col-md-6\">\n"
-    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.requests : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large","containerClass":"metric-container col-md-2"},"data":data,"indent":"  ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
-    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.success : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large success-metric","containerClass":"success-metric-container metric-container col-md-2"},"data":data,"indent":"  ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
-    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.failures : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large failure-metric","containerClass":"failure-metric-container metric-container col-md-2"},"data":data,"indent":"  ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
+  return "<div class=\"client-metrics row\">\n  <div class=\"col-md-2\">\n"
+    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.success : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large success-metric","containerClass":"success-metric-container metric-container col-md-2"},"data":data,"indent":"    ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
+    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.requests : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large","containerClass":"metric-container col-md-2"},"data":data,"indent":"    ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
+    + "  </div>\n\n  <div class=\"col-md-2\">\n"
+    + ((stack1 = container.invokePartial(partials.rateMetricPartial,((stack1 = (depth0 != null ? depth0.metrics : depth0)) != null ? stack1.failures : stack1),{"name":"rateMetricPartial","hash":{"metricClass":"metric-large failure-metric","containerClass":"failure-metric-container metric-container col-md-2"},"data":data,"indent":"    ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
+    + "  </div>\n\n"
+    + ((stack1 = container.invokePartial(partials.latencyPartial,(depth0 != null ? depth0.latencies : depth0),{"name":"latencyPartial","data":data,"indent":"  ","helpers":helpers,"partials":partials,"decorators":container.decorators})) != null ? stack1 : "")
     + "</div>\n";
 },"usePartial":true,"useData":true});
 templates['router_server_container'] = template({"compiler":[7,">= 4.0.0"],"main":function(container,depth0,helpers,partials,data) {
-    return "<div class=\"router-server clearfix\">\n  <div class=\"server-metrics\"></div>\n  <div class=\"server-success-chart col-md-6\">\n    <div class=\"router-graph-header\">Server success rate</div>\n  </div>\n</div>\n";
+    var helper;
+
+  return "<div class=\"server-header router-header-large\">\n  "
+    + container.escapeExpression(((helper = (helper = helpers.server || (depth0 != null ? depth0.server : depth0)) != null ? helper : helpers.helperMissing),(typeof helper === "function" ? helper.call(depth0 != null ? depth0 : {},{"name":"server","hash":{},"data":data}) : helper)))
+    + "\n</div>\n\n<div class=\"router-server clearfix\">\n  <div class=\"server-metrics metrics-container col-md-6\"></div>\n  <div class=\"server-success-chart col-md-6\">\n    <div class=\"router-graph-header\">Server success rate</div>\n  </div>\n</div>\n";
 },"useData":true});
 templates['router_summary'] = template({"1":function(container,depth0,helpers,partials,data) {
     var helper;