+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/H2Config.scala
@@ -7,12 +7,14 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.{DeserializationContext, JsonDeserializer, JsonNode}
 import com.twitter.conversions.storage._
 import com.twitter.finagle.buoyant.PathMatcher
-import com.twitter.finagle.buoyant.h2.{LinkerdHeaders, Request, Response, ResponseClassifiers}
+import com.twitter.finagle.buoyant.h2._
 import com.twitter.finagle.buoyant.h2.param._
 import com.twitter.finagle.client.StackClient
-import com.twitter.finagle.{Path, Stack, param}
+import com.twitter.finagle.netty4.ssl.server.Netty4ServerEngineFactory
+import com.twitter.finagle.{Stack, param}
 import com.twitter.util.Monitor
 import io.buoyant.config.PolymorphicConfig
+import io.buoyant.router.h2.DupRequest
 import io.buoyant.router.{ClassifiedRetries, H2, RoutingFactory}
 import io.netty.handler.ssl.ApplicationProtocolNames
 import scala.collection.JavaConverters._
@@ -30,7 +32,7 @@ class H2Initializer extends ProtocolInitializer.Simple {
 
     // retries can't share header mutations
     val pathStack = H2.router.pathStack
-      .insertAfter(ClassifiedRetries.role, h2.DupRequest.module)
+      .insertAfter(ClassifiedRetries.role, DupRequest.module)
       .prepend(LinkerdHeaders.Dst.PathFilter.module)
 
     // I think we can safely ignore the DelayedRelease module (as
@@ -43,9 +45,9 @@ class H2Initializer extends ProtocolInitializer.Simple {
       .prepend(LinkerdHeaders.Dst.BoundFilter.module)
 
     val clientStack = H2.router.clientStack
+      .replace(H2TraceInitializer.role, H2TraceInitializer.clientModule)
       .insertAfter(StackClient.Role.prepConn, LinkerdHeaders.Ctx.clientModule)
 
-    //  .replace(HttpTraceInitializer.role, HttpTraceInitializer.clientModule)
     //  .insertAfter(Retries.Role, http.StatusCodeStatsFilter.module)
 
     H2.router
@@ -56,10 +58,15 @@ class H2Initializer extends ProtocolInitializer.Simple {
 
   private[this] val monitor = Monitor.mk { case NonFatal(_) => true }
 
-  protected val defaultServer = H2.server.withStack(H2.server.stack
-    .prepend(LinkerdHeaders.Ctx.serverModule)
-    .prepend(h2.ErrorReseter.module))
-    .configured(param.Monitor(monitor))
+  protected val defaultServer = {
+    val stk = H2.server.stack
+      .replace(H2TraceInitializer.role, H2TraceInitializer.serverModule)
+      .prepend(LinkerdHeaders.Ctx.serverModule)
+      .prepend(h2.ErrorReseter.module)
+
+    H2.server.withStack(stk)
+      .configured(param.Monitor(monitor))
+  }
 
   override def clearServerContext(stk: ServerStack): ServerStack = {
     // Does NOT use the ClearContext module that forcibly clears the
@@ -145,8 +152,8 @@ class H2PrefixConfig(prefix: PathMatcher) extends PrefixConfig(prefix) with H2Cl
 trait H2ClientConfig extends ClientConfig with H2EndpointConfig {
 
   @JsonIgnore
-  override def params(vars: Map[String, String]) = withEndpointParams(super.params(vars))
-
+  override def params(vars: Map[String, String]): Stack.Params =
+    withEndpointParams(super.params(vars))
 }
 
 @JsonTypeInfo(
@@ -189,6 +196,9 @@ class H2ServerConfig extends ServerConfig with H2EndpointConfig {
   @JsonIgnore
   override val alpnProtocols: Option[Seq[String]] =
     Some(Seq(ApplicationProtocolNames.HTTP_2))
+
+  @JsonIgnore
+  override val sslServerEngine = Netty4ServerEngineFactory()
 
   override def withEndpointParams(params: Stack.Params): Stack.Params = super.withEndpointParams(params)
     .maybeWith(maxConcurrentStreamsPerConnection.map(c => Settings.MaxConcurrentStreams(Some(c.toLong))))